feat(console): move init storage heap state to cdn
