+++ src/console/src/lifecycle.rs
@@ -1,9 +1,10 @@
 use crate::cdn::certified_assets::upgrade::defer_init_certified_assets;
-use crate::memory::{get_memory_upgrades, init_stable_state, init_storage_heap_state, STATE};
+use crate::memory::{get_memory_upgrades, init_stable_state, STATE};
 use crate::types::state::{Fees, HeapState, Rates, ReleasesMetadata, State};
 use ciborium::{from_reader, into_writer};
 use ic_cdk::caller;
 use ic_cdk_macros::{init, post_upgrade, pre_upgrade};
+use junobuild_cdn::lifecycle::init_cdn_storage_heap_state;
 use junobuild_shared::controllers::init_controllers;
 use junobuild_shared::upgrade::{read_post_upgrade, write_pre_upgrade};
 use std::collections::HashMap;
@@ -19,7 +20,7 @@ fn init() {
         controllers: init_controllers(&[manager]),
         rates: Rates::default(),
         fees: Fees::default(),
-        storage: init_storage_heap_state(),
+        storage: init_cdn_storage_heap_state(),
         releases_metadata: ReleasesMetadata::default(),
     };
