+++ src/libs/cdn/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod constants;
+pub mod lifecycle;
 pub mod proposals;
 pub mod storage;
 pub mod strategies;
